backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
mysqlclient==2.2.0
pymysql==1.1.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
python-dotenv==1.0.0
Werkzeug==2.3.7
firebase-admin==6.2.0
gunicorn==21.2.0 